Reduce FOR memory usage
FOR allocates a large buffer to perform its variable substitution,
resizing it once done.  This is done on each loop, with all buffers
being freed when the FOR finishes.  It's not normally an issue, but
our RANGE feature could potentially be a very large loop, resulting in
vast amounts of memory being used.  Allocate the buffer once, reusing it
on each loop.  There are also some other smaller allocations made, so
free those on each loop, too.