Allow load via RegSvr32
Allow RegSvr32 to load Enhanced Batch.  This requires the "/e" option
(to prevent the succeeded message) or "/s" option (to prevent all
messages).  If the DLL itself fails to load ERRORLEVEL will be 3,
otherwise it will be 0 (whether it loaded or not).

The RunDLL command may end with "/s" to suppress messages from Enhanced
Batch (failing to find/load the DLL will still show a message).