Add FOR shortcuts
Allow briefer FOR statements:

    for do ...          ->  for %%_ in (:*) do ...
    for %%? do ...      ->  for %%? in (:range*) do ...
    for %%? N ...       ->  for %%? in (:range*N) ...

There must be whitespace after "do", the other delimiters ("=,;") will
not be recognised.  There must only be a single whitespace after the
variable for "do" or the number to be recognised.  A delayed variable is
assumed to be a number, use two spaces if that is not the case.

FOR /F has a new option "line", which sets "delims" and "eol" to NUL,
ensuring a complete line will be read (assuming "tokens" is still 1).