Improve finding labels
If a label is only ASCII search the batch file directly.  It must start
the line for the search to be successful.  If it starts with '~' the
search will be performed backwards, continuing from the end if it's not
found at the start (e.g.: `goto ~loop` will search for the previous
`:loop`).