Return nothing for unknown modifier on the CLI
An unknown modifier would expand to the original variable name, to make
it easy to know something's wrong.  However, this broke FOR expansion on
the command line, where something like "%I;%I" would expand to "I;I"
(variable I with missing modifier expanded to "I;").  Return nothing on
the command line so this works as expected.