Add CLEAR call extension
Erase the entire window (not the buffer) or any rectangle within it.  If
the first call does not specify a rectangle the current row will become
the top line of a new window (thus preserving existing lines, as much as
the buffer allows).  The cursor moves to the top-left of the rectangle;
there is an option to prevent that.  Options allow specifying the
attributes and/or character to use.  There is also a "fill" option,
which sets just the attributes or character, preserving the other.