Allow a single image to contain multiple frames
Add options to IMAGE to treat an image as a sprite sheet.  Use /FC to
count the number of sprites (automatically determining the size), or /FS
to define the size of each sprite (automatically determining the count).
There are also options to skip pixels between the sprites (ignore an
explicit grid).