Fix delayed CmdBatNotificationStub
If EB was loaded from the Vista+ command line before any batch files
were run the next batch would crash.  This was due to hooking the
delay import code that loaded the Stub, so the hook would end up
calling itself.  In this case bypass the delay load function and point
directly to the notification function.