Add ':=' for implicit expansion
Values starting with '@' would automatically be expanded when set (to
make it look a little nicer and avoid requiring delayed expansion).
However, this caused problems with SET /P (the answer would also be
expanded) and creating variables that should start with '@'.  Create
':=' as an expression assignment, which expands its value.  Currently,
if the value doesn't exist as a variable it is set as usual, but that
will change in the future.