Add TRACE variable extension
Trace is an alternative to ECHO ON.  It will write commands to standard
error, using the BATCHFILE setting.  Commands are written after delayed
expansion.  Only directly executed commands are included, not
"subsidiary" commands (e.g. given "if COND (echo True) else echo False"
only one of the echos will be written, not the IF as a whole).  To help
distinguish commands from output when writing to the console, commands
are enclosed in square brackets (e.g. "test.bat:5: [rem]"); this does
not occur when stderr is redirected to file (e.g. "test.bat:5: rem").