Add modifiers to work with keys
KEY will extract the name of the key; SHIFT, CTRL and ALT will test if
the appropriate modifiers are present (returning "0" or "1"); using both
will return "0" if the modifiers are not present, the appropriate key if
so.  If the variable is not a key "0" is returned.  Due to the way these
modifiers are processed they are expected to be used by themselves;
combining with other modifiers is undefined.