Reduce ELEVATE security hole
The elevated process would remain active for 15 minutes, potentially
allowing any other process to wait for it and use it.  Create a key
based on the PID and address of the batch file, making it a little
harder to hijack.  Elevation will now remain active for the remainder of
the batch.  The prompt details will show the title of the CMD window
requesting elevation.